boost_control: Correct type of cpubw_boost_freq
* ../drivers/devfreq/devfreq_boost.c:144:11: warning: comparison of distinct pointer types ('typeof (cpubw_boost_freq) *' (aka 'unsigned int *') and 'typeof (df->max_freq) *' (aka 'unsigned long *'))
      [-Wcompare-distinct-pointer-types]
                               min(cpubw_boost_freq, df->max_freq) :
                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../include/linux/kernel.h:734:17: note: expanded from macro 'min'
        (void) (&_min1 == &_min2);              \
                ~~~~~~ ^  ~~~~~~

+++ drivers/cpufreq/boost_control.c
@@ -25,8 +25,8 @@ module_param_named(input_boost_freq_lp, input_boost_freq_lp, int, 0644);
 u32 wake_boost_duration=CONFIG_WAKE_BOOST_DURATION_MS;
 module_param_named(cpu_wake_boost_duration, wake_boost_duration, int, 0644);
 
-u32 cpubw_boost_freq=CONFIG_DEVFREQ_MSM_CPUBW_BOOST_FREQ;
-module_param_named(cpubw_boost_freq, cpubw_boost_freq, int, 0644);
+unsigned long cpubw_boost_freq=CONFIG_DEVFREQ_MSM_CPUBW_BOOST_FREQ;
+module_param_named(cpubw_boost_freq, cpubw_boost_freq, long, 0644);
 
 u32 cpubw_wake_boost_duration=CONFIG_DEVFREQ_WAKE_BOOST_DURATION_MS;
 module_param_named(cpubw_wake_boost_duration, cpubw_wake_boost_duration, int, 0644);
